% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interactive_weather_plot.R
\name{interactive_weather_plot}
\alias{interactive_weather_plot}
\title{Weather forecast application}
\usage{
interactive_weather_plot()
}
\value{
A shiny application
}
\description{
A Shiny application based on get_weather_forecast() from the Weather4cat package
}
