% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack_disc.R
\docType{package}
\name{ShinyWeatherApp}
\alias{ShinyWeatherApp}
\title{Shiny app for visualizing weather data forecasts.}
\description{
An interactive plot generator.
}
\author{
Leonard Persson Norblad \email{leope744@student.liu.se} Jacob Sibanda \email{jacsi259@student.liu.se}
}
